requirements.txt
diffusers
transformers
scipy
ftfy
accelerate
datasets
torchvision
wandb
xformers
bitsandbytes
rtpt
torchmetrics
git+https://github.com/openai/CLIP.git
git+https://github.com/modestyachts/ImageNetV2_pytorch
einops
